recommendations/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.1
pydantic==2.4.2
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.2.2 